pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "python-decouple>=3.8",
]
